Implement the Vigenere Cipher.
Add a class for the Vigenere cipher with two functions.
One for setting up the key (only uppercase letters) and
a map containing the key letter and a corresponding Caesar cipher
object with key the position of the letter in the alphabet.
The second applies the Vigenere cipher, by applying the Caesar cipher
(of the previously set map, i.e. with key the position of the key letter)
for each input letter separately.